+++ src/main/java/com.outsystems.plugins.inappbrowser/osinappbrowserlib/helpers/OSIABCustomTabsSessionHelper.kt
@@ -47,7 +47,6 @@ class OSIABCustomTabsSessionHelper: OSIABCustomTabsSessionHelperInterface {
                 }
 
                 override fun onServiceDisconnected(name: ComponentName) {
-                    context.unbindService(this)
                     customTabsSessionCallback(null)
                 }
             }
@@ -61,6 +60,8 @@ class OSIABCustomTabsSessionHelper: OSIABCustomTabsSessionHelperInterface {
     ) : CustomTabsCallback() {
 
         private var isCustomTabsActivityOnTop = false
+        private var pendingTabHiddenEvent = false
+
         init {
             var browserEventsJob: Job? = null
 
@@ -69,9 +70,16 @@ class OSIABCustomTabsSessionHelper: OSIABCustomTabsSessionHelperInterface {
                     when (event.action) {
                         OSIABCustomTabsControllerActivity.EVENT_CUSTOM_TABS_RESUMED -> {
                             isCustomTabsActivityOnTop = true
+                            if (pendingTabHiddenEvent) {
+                                pendingTabHiddenEvent = false
+                                lifecycleScope.launch {
+                                    OSIABEvents.postEvent(OSIABEvents.BrowserFinished(browserId))
+                                }
+                            }
                         }
                         OSIABCustomTabsControllerActivity.EVENT_CUSTOM_TABS_PAUSED -> {
                             isCustomTabsActivityOnTop = false
+                            pendingTabHiddenEvent = false
                         }
                         OSIABCustomTabsControllerActivity.EVENT_CUSTOM_TABS_DESTROYED -> {
                             browserEventsJob?.cancel()
@@ -90,6 +98,7 @@ class OSIABCustomTabsSessionHelper: OSIABCustomTabsSessionHelperInterface {
                     }
                     else {
                         // App not open but custom tabs is hidden (home button, recent apps, etc.)
+                        pendingTabHiddenEvent = true
                         return
                     }
                 }